[BUGFIX] changed "eq" to "equalp"
From: http://www.cs.cmu.edu/Groups/AI/html/cltl/clm/node74.html
== eq ==
(eq x y) is true if and only if x and y are the same identical object.
(Implementationally, x and y are usually eq if and only if they address
the same identical memory location.)

== equalp ==
Two objects are equalp if they are equal; if they are characters and
satisfy char-equal, which ignores alphabetic case and certain other
attributes of characters; if they are numbers and have the same numerical
value, even if they are of different types; or if they have components
that are all equalp.